Updated the Chromosome data structure to use Gene pointers instead of managed Gene variables (to allow for heterogeneous chromosomes in the future). Updated the Gene data structure to allow for per-gene dynamic mutation rates; if enabled for an arbitrary gene, at each mutation, that gene can change how frequently it is selected for mutation.
Updated the fitness function in AudioGenome to take into account the
duration of notes played so as to minimize the amount of silence per
song and maximize the amount of non-silence.

The next step is to implement the new model detailed in the readme,
most importantly the rhythm data structure, to make the generated
songs sound more natural and less random.